Separating redux parts and giving redux a dedicated folder
